(general) fixing emitted warnings

+++ src/Basescape/BasescapeState.cpp
@@ -226,7 +226,7 @@ void BasescapeState::init()
 	}
 
 	_txtFunds->setText(tr("STR_FUNDS").arg(Unicode::formatFunding(_game->getSavedGame()->getFunds())));
-	_btnNewBase->setVisible(_game->getSavedGame()->getBases()->size() < Options::maxNumberOfBases);
+	_btnNewBase->setVisible(static_cast<int>(_game->getSavedGame()->getBases()->size()) < Options::maxNumberOfBases);
 
 	if (!_game->getMod()->getNewBaseUnlockResearch().empty())
 	{
